Move 'is-visible' to private folder
