+++ addon/create.js
@@ -1,7 +1,7 @@
 import Ceibo from 'ceibo';
 import { text } from './queries/text';
-import { isVisible } from './predicates/is-visible';
-import { isHidden } from './predicates/is-hidden';
+import { isVisible } from './-private/properties/is-visible';
+import { isHidden } from './-private/properties/is-hidden';
 import { contains } from './-private/properties/contains';
 import { clickOnText } from './-private/properties/click-on-text';
 import { clickable } from './-private/properties/clickable';
